Fixed Documentation Errors in S and T gate
The S gate which applies a PI/2 phase shift and T gate which applies a PI/4 phase shift gate were wrongly documented to term them as PI/4 and PI/8 gate respectively. The PI/8 gate is a valid and used name for T gate but it is quite confusing for beginners since the matrix only adds in a phase of PI/4 if qubit is 1.