Refactor target packages to use interfaces (#447)
This change refactors the target packages to define interfaces for the natively supported Q# callables, which provides an important layer of indirection that lets the Q#/C# in the target package call into those interfaces while the simulator implements those interfaces, all without either taking a direct reference on the other. This means I can once again unify the simulator into a single package, where it dynamically detects the target package intrinsics to override at runtime via reflection.

Fixes #432

Major changes in this package:

Introduce Simulation/TargetDefinition/Interfaces that defines the package with the union of all interfaces for all target packages and target specific interfaces that collect the set of gates supported by that target (expected to eventually be auto-generated from Q# definitions).

Implementations of the callables in Simulation/TargetDefinition/Intrinsic that implement those abstract classes and call into the interfaces (also expected to be auto-generated, perhaps replacing the abstract classes altogether).

Adapted the full state QuantumSimulator, ToffoliSimulator, and ResourceEstimator to use the interfaces such that the single simulator package can now be used with all three target packages (simulator no longer has compile time dependency on the target package). This means Simlators.Impl and Simulators could be combined back into a single folder, and Simulators.Type1 and Simulators.Type2 eliminated.

Updated test project definitions to include references to both the simulator and the desired target package, matching the pattern we expect callers to use.

Moved the QuantumProcessor implementation into the QsharpCore package so that this implementation can remain the same. Converting it to the interface based approach would be a breaking change, but to keep it the same required putting it in a place that could safely compile against both QsharpCore and the SimulatorBase class. For the moment that means that the QsharpCore target package has a compile time dependency on the Simulators package, which doesn't match the design intent but doesn't cause any problems. Eventually, QuantumProcessor can either be moved to its own package or removed entirely in favor of broader adoption of the interface based design.

Updates to C# generation that provide an opt-in mechanism for intrinsic Q# callables to be generated as concrete classes instead of abstract, and the generated code calls into the interface classes defined as part of 1 (thanks @ScottCarda-MS). This means the intrinsic callables that are part of the target package no longer need to define a child class that registers as an override of the abstract parents.