Add Type3 target package
This adds the Type3 target package, which is very similar to Type1 but includes support for an intrinsinc native SWAP, and uses a NOOP when handing post-measurement preparation (reset and re-prepare of qubit state does not need to be explicit). This includes the package itself, simlator support, tests, and updates to the solution and packing script.