Fix dynamic assembly load logic in Tracer
This fixes an issue where the fallback logic in the `QCTraceSimulator` for dynamically loading the QSharp.Core target package wouldn't work for manually written C# drivers. This fixes the issue by introducing new fallback logic that relies on the assembly name of the current simulator assembly to determine the full name of the QSharp.Core assembly to load.

Fixes #540
Fixes #542