Remove BitStates.hpp (#581)
* Remove BitStates.hpp

BitStates.hpp defined a custom implementation for tracking vectors of bits. However, `vector<bool>` from the STL already is a specialization of `vector` designed to save space and use a single bit for each index. Since we don't want to support the custom type in our API nor do we really need one internally, it was removed and all instances replaced with `vector<bool>`.

Fixes #577

* CR feedback