Update QIR tests for latest calling conventions (#667)
This change updates the calling conventions used in the QIR Runtime tests to use `__Interop` wrappers for calling functions. This also required updating some older tests to the multiple entry point strategy for testing so that each test function is considered an externally callable operation/function. In addition, this updates the CMakeLists.txt to consume the bitcode files instead of the human readable .ll files wherever possible.