Introduce ClassicalControl namespace, depreicate old namespace
This moves the implementation of the callables used by the Classical Control rewrite step into a new namespace, Microsoft.Quantum.ClassicalControl. The old callables in the Microsoft.Quantum.Simulation.QuantumProcessor.Extensions are marked as depricated and will be removed in a later PR. Follow up will need to be done across the remaining repos where explicit references to this namespace exist (such as in the compiler rewrite step itself) to move to the new namespace.