Remove CSharpGeneration package
This change removes the CSharpGeneration package in favor of having it built from within compiler (see https://github.com/microsoft/qsharp-compiler/pull/1054). This is part of resolving #618. This change is not expected to build properly until the PR to qsharp-compiler is completed.

Also of note, this removes the old Microsoft.Quantum.Development.Kit which is no longer used and all of the issues in the readme there with custom .NET SDKs have been resolved.