Update QIR Runtime to use QubitIdType
This updates the QIR Runtime to use a consistent type for qubit ids, `typedef intptr_t QubitIdType;`. All APIs that correspond to the QIR spec continue to use opaque pointer type `QUBIT*` and then immediately cast to `QubitIdType` for use internally. Of note, the Qubit Manager uses negative values for special meaning when tracking qubits in use, so the unsigned `uintptr_t` is not compatible.