Update QIR Runtime to use QubitIdType (#833)
* Update QIR Runtime to use QubitIdType

This updates the QIR Runtime to use a consistent type for qubit ids, `typedef intptr_t QubitIdType;`. All APIs that correspond to the QIR spec continue to use opaque pointer type `QUBIT*` and then immediately cast to `QubitIdType` for use internally. Of note, the Qubit Manager uses negative values for special meaning when tracking qubits in use, so the unsigned `uintptr_t` is not compatible.

* Add [[maybe_unused]] for Toffoli

* Fix test files

* Fix Qubit Manager assert

* Update src/Qir/Runtime/lib/Simulators/FullstateSimulator.cpp

Co-authored-by: Robin Kuzmin <9372582+kuzminrobin@users.noreply.github.com>

* Update src/Qir/Runtime/unittests/QirRuntimeTests.cpp

Co-authored-by: Robin Kuzmin <9372582+kuzminrobin@users.noreply.github.com>

* Update src/Qir/Runtime/unittests/QirRuntimeTests.cpp

Co-authored-by: Robin Kuzmin <9372582+kuzminrobin@users.noreply.github.com>

* Fix alignment

* Update comment description

Co-authored-by: Robin Kuzmin <9372582+kuzminrobin@users.noreply.github.com>