Update native simulator APIs to avoid locking
We've had on-again, off-again issues with the maps used by the native simulators due to difficult to track down synchronization bugs. There intermittent failures would persist even though the global maps from simulator id to simulator pointer seem to be synchronized by reader-writer locks appropriately. This change instead removes the maps and the locks entirely in favor of returing the simulator pointer itself as the id, such that when the caller passes it in to invoke a method we can simply cast it to the required simulator type. Note that this is a breaking change to the native API as the width of the simulator id has changed from `unsigend` to `uintptr_t`.