Refactoring decompositions for hardware patterns
This change refactors the decompositions to avoid specific patterns. Namely recursion, array concatenation and slicing, and complex mathematical operations. It uses the new decomposition tests to validate the refactor across controlled and adjoint specializations with up to 8 controls. This should make the decompositions more compatible with future hardware targeting transformations.

Note that this updatest to a beta version of the QDK to pick up a compiler fix that is required to compile the decompositions.