Add Type4 target package, update `Measure` decomp
This change adds a Type4 package that avoids qubit reuse, and updates the Type2 package to use similar patterns. It adds tests, packaging and publishing for Type4.

In addition, this updates the decomposition of `Measure` in terms of single qubit Pauli-Z measurement to be limited to 8 qubits, updating all the target packages in the process.