Remove unused Rust sys crate, minor clean up
This removes the Rust sys crate that wrapped the C++ runtime since it was no longer used. In addition, it moves some rustc flags configuration out of build scripts and into a repo-wide settings file, resolves a cargo warning by moving profile settings up to the root of the workspace, and cleans up a few leftover functions listed as dllexport in the runtime header files despite having no implementation.